removed some debug bind lob length was a short on client, but a long on server, fixed that

+++ src/api/c++/include/sqlrelay/private/bindvar.h
@@ -29,7 +29,7 @@ class bindvar {
 			char	*lobval;
 			short	cursorid;
 		} value;
-		unsigned short	valuesize;
+		unsigned long	valuesize;
 		bindtype	type;
 		unsigned short	send;
